# Simplify the lru implementation