--- a/enterprise/server/remote_execution/filecache/filecache.go
+++ b/enterprise/server/remote_execution/filecache/filecache.go
@@ -77,7 +77,7 @@ var (
 type fileCache struct {
 	rootDir     string
 	lock        sync.RWMutex
-	l           interfaces.LRU[*entry]
+	l           *lru.LRU[*entry]
 	dirScanDone chan struct{}
 }