Now that keys are always strings since https://github.com/buildbuddy-io/buildbuddy/pull/4379, we can just use a regular map with string keys, instead of computing hashes manually.

I also added a filecache benchmark which shows a minor improvement in CPU and memory.
```
goos: linux
goarch: amd64
cpu: AMD Ryzen 9 7900X 12-Core Processor
                            │ /tmp/baseline │          /tmp/simplekeys           │
                            │    sec/op     │   sec/op     vs base               │
ContainsAdd/100%Added/1K-24     15.36m ± 1%   15.30m ± 2%  -0.42% (p=0.012 n=19)
ContainsAdd/95%Added/1K-24      15.38m ± 1%   15.24m ± 1%  -0.91% (p=0.005 n=19)
ContainsAdd/50%Added/1K-24      15.44m ± 0%   15.24m ± 1%  -1.28% (p=0.001 n=19)
ContainsAdd/0%Added/1K-24       15.41m ± 1%   15.30m ± 1%       ~ (p=0.116 n=19)
geomean                         15.40m        15.27m       -0.83%

                            │ /tmp/baseline │           /tmp/simplekeys           │
                            │     B/op      │     B/op      vs base               │
ContainsAdd/100%Added/1K-24    1.386Mi ± 1%   1.378Mi ± 1%  -0.54% (p=0.013 n=19)
ContainsAdd/95%Added/1K-24     1.379Mi ± 1%   1.374Mi ± 1%       ~ (p=0.212 n=19)
ContainsAdd/50%Added/1K-24     1.386Mi ± 0%   1.371Mi ± 1%  -1.08% (p=0.002 n=19)
ContainsAdd/0%Added/1K-24      1.382Mi ± 0%   1.377Mi ± 1%       ~ (p=0.057 n=19)
geomean                        1.383Mi        1.375Mi       -0.61%

                            │ /tmp/baseline │          /tmp/simplekeys           │
                            │   allocs/op   │  allocs/op   vs base               │
ContainsAdd/100%Added/1K-24     17.34k ± 1%   16.38k ± 1%  -5.51% (p=0.000 n=19)
ContainsAdd/95%Added/1K-24      17.26k ± 1%   16.33k ± 0%  -5.40% (p=0.000 n=19)
ContainsAdd/50%Added/1K-24      17.34k ± 0%   16.30k ± 1%  -5.97% (p=0.000 n=19)
ContainsAdd/0%Added/1K-24       17.29k ± 0%   16.36k ± 1%  -5.38% (p=0.000 n=19)
geomean                         17.31k        16.35k       -5.57%
```