@vadimberezniker 2025-02-14
Do you have a rough idea of what the memory increase is going to be? IIRC the reason we added the hashing was to reduce memory usage of the disk cache which used file paths as keys. The file cache has shorter keys so it's likely much less of an issue but it's worth measuring what the memory impact is of storing the larger keys.