@vanja-p 2025-02-14
Hmm, I didn't consider the size of the keys. Do you have any ideas on how I should check the impact? I'll try to come up with something either way.

Do you think I should roll this back in the meantime?