@vadimberezniker 2025-02-14
> Hmm, I didn't consider the size of the keys. Do you have any ideas on how I should check the impact? I'll try to come up with something either way.
> 
> Do you think I should roll this back in the meantime?

I don't think the file cache keys are long enough that it should warrant a rollback. 

I realize my earlier comment regarding disk cache wasn't clear enough. This would have the biggest impact on disk cache users but most people have moved on from disk cache to pebble cache so that's not something to worry about today.

I can't think of a simple way to do this via a benchmark, but if you have a full local filecache already you can start it with and without this change and look at the heap profile to see what the difference in memory usage is.