@vanja-p 2025-02-14
I'll try the memory profile approach. Note that this change affects all uses of the LRU, not just the file cache.