@vadimberezniker 2025-02-14
> I'll try the memory profile approach. Note that this change affects all uses of the LRU, not just the file cache.

I singled out the file cache because I expect it to be the largest user of the lru (by number of entries), behind the disk cache.