@vanja-p 2025-02-16
I did some benchmarks and took some profiles and this change makes the LRU heap 10%-20% bigger, for both the filecache and the disk_cache.

Based on profiles of prod executors using lots of RAM, it seems like the filecache LRU currently uses 20% of the total RAM, so this PR will increase executor memory usage by 4%, but I have a pretty straightforward way to fix that and actually reduce memory usage compared to the old LRU. See https://github.com/buildbuddy-io/buildbuddy/pull/8420.

For the disk cache, I'm not sure how much of the app's memory the LRU uses, since it's not used in prod. I ran a local `//server` with a 10GB disk cache, and the LRU seemed to use only 5% of the memory.

Let me know what you think, but unless the current impact is too severe or uncertain, I think this can stay in for now, and I'll push the filecache optimization soon.